+++ app/content/tasks/event.py
@@ -5,7 +5,13 @@ from django.conf import settings
 from sentry_sdk import capture_exception
 
 from app.celery import app
+from app.common.enums import AdminGroup
 from app.communication.notifier import Notify
+from app.communication.slack import Slack
+from app.constants import (
+    SLACK_EVENTS_NOK_CHANNEL_ID,
+    SLACK_EVENTS_OTHER_CHANNEL_ID,
+)
 from app.content.models.strike import create_strike
 from app.util.mail_creator import MailCreator
 from app.util.tasks import BaseTask
@@ -69,7 +75,9 @@ def run_sign_up_start_notifier(self, *args, **kwargs):
         for event in events:
             __sign_up_start_notifier(event)
 
-        self.logger.info(f'Runned "run_sign_up_start_notifier" for {events.count()} events')
+        self.logger.info(
+            f'Runned "run_sign_up_start_notifier" for {events.count()} events'
+        )
     except Exception as e:
         capture_exception(e)
 
@@ -155,25 +163,25 @@ def __post_event_actions(event, *args, **kwargs):
 
 
 def __sign_up_start_notifier(event, *args, **kwargs):
-    from app.content.models import User
-
-    users = User.objects.filter(
-        sign_up_start_notifications=True
+    CHANNEL_ID = (
+        SLACK_EVENTS_NOK_CHANNEL_ID
+        if event.organizer
+        and str(event.organizer.slug).lower() == str(AdminGroup.PROMO).lower()
+        else SLACK_EVENTS_OTHER_CHANNEL_ID
     )
-    description = [
-        f"Påmeldingen for {event.title} har nå åpnet! Påmeldingen er åpen frem til {datetime_format(event.end_registration_at)}, men husk at det kan bli fullt før den tid.",
-        "PS: Om du ikke lenger ønsker å motta påminnelse om åpning av påmelding til arrangementer kan dette skrus av i profilen din.",
-    ]
-    Notify(users, f"Påmelding for {event.title} har åpnet").send_email(
-        MailCreator(f"Påmelding for {event.title} har åpnet")
-        .add_paragraph("Hei!")
-        .add_paragraph(description[0])
-        .add_paragraph(description[1])
-        .add_event_button(event.id)
-        .generate_string(),
-    ).send_notification(
-        description=" \n".join(description), link=event.website_url
+    slack = (
+        Slack(
+            channel_id=CHANNEL_ID,
+            fallback_text=f'Påmelding til "{event.title}" har nå åpnet!',
+        )
+        .add_header(event.title)
+        .add_markdwn(
+            f'Påmelding til "{event.title}" har nå åpnet! 🏃 Arrangementet starter {datetime_format(event.start_date)} og har {event.limit} plasser. Påmeldingen er åpen frem til {datetime_format(event.end_registration_at)}, men husk at det kan bli fullt før det. ⏲️\n\n<{settings.WEBSITE_URL}{event.website_url}|*Se arrangementet her og meld deg på nå!*>'
+        )
     )
+    if event.image:
+        slack.add_image(event.image, event.image_alt or event.title)
+    slack.send()
 
     event.runned_sign_up_start_notifier = True
     event.save(update_fields=["runned_sign_up_start_notifier"])