+++ CHANGELOG.md
@@ -13,7 +13,7 @@
 ---
 
 ## Neste versjon
-- ✨ **Arrangementer**. Det er mulig å skru på påminnelser om at påmelding til arrangementer har åpnet.
+- ✨ **Arrangementer**. Varsel om at påmelding har åpnet sendes nå i Slack.
 
 ## Versjon 2022.03.05
 - 🦟 **Bilder**. Fikset feil som oppstod ved opplastning av profilbilde.